libf2fs_zoned: factor our helper to get chunk sectors
This moves the code which fetches the chunk_sectors into a helper.
Yes, this could in theory be used by non-zoned devices but that's
not the case yet, so no need to make this a generic routine.

This makes it clear what this is doing, and helps us make the
subsequent changes easier to read.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>