libf2fs: add support to report zone size
Expand get_device_info() to report the zone size.
This is now important give power of 2 zone sizees (PO2)
can exist, and so can non power of 2 zones sizes (NPO2),
and we should be aware of the differences in terms of
support.

This will be used more in subsequent patch.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>