libf2fs: don't allow mkfs / fsck on zoned NPO2
f2fs currently only work with zoned storage devices with a zone
size which is a power of 2 (PO2). So check if a non-power of 2
zone is device is found, and if so disallow its use. This prevents
users from incorrectly using these devices.

This is a non-issue today give today's kernel does not allow NPO2
zon devices to exist. But these devices do exist, and so proactively
put a stop-gap measure in place to prevent the from being assumed
to be used.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>